E2E: Che integration tests updated with new script

+++ ee_tests/local_cico_run_e2e_tests.sh
@@ -14,7 +14,7 @@ rm -rf target
 mkdir -p dist target
 
 # Shutdown container if running
-if [ -n "$(docker ps -q -f name=fabric8-test)" ]; then
+if [ -n "$(docker ps -a -q -f name=fabric8-test)" ]; then
     docker rm -f fabric8-test
 fi
 
@@ -24,7 +24,8 @@ docker run --shm-size=256m --detach=true --name=fabric8-test --cap-add=SYS_ADMIN
           -e OSO_USERNAME -e GITHUB_USERNAME -e GITHUB_REPO -e TEST_SUITE -e QUICKSTART_NAME -e RELEASE_STRATEGY \
           -e FEATURE_LEVEL -e RESET_ENVIRONMENT -e DEBUG \
           -e "FORCE_COLOR=1" \
-          -t -v "$(pwd)/dist:/dist:Z" fabric8-test:latest
+          -t -v "$(pwd)/dist:/dist:Z" -v /var/run/docker.sock:/var/run/docker.sock \
+          fabric8-test:latest
 
 # Start Xvfb
 docker exec fabric8-test /usr/bin/Xvfb :99 -screen 0 1024x768x24 &
@@ -36,7 +37,7 @@ docker exec fabric8-test ./ts-protractor.sh "$TEST_SUITE" | tee target/protracto
 docker cp fabric8-test:/opt/fabric8-test/target/. target
 
 # Shutdown container if running
-if [ -n "$(docker ps -q -f name=fabric8-test)" ]; then
+if [ -n "$(docker ps -a -q -f name=fabric8-test)" ]; then
     docker rm -f fabric8-test
 fi
